Change the license of JUCE.cmake from MIT to GPLv3
Reprojucer.cmake and the CMakeLists.txt files for the JUCE examples and
extras are not directly using JUCE, so they don't have to be licensed
under the terms of the GNU General Public License. However, it is
easier to deal with only one license in JUCE.cmake, so let's unify and
use GPLv3 for whole of JUCE.cmake.