Do not rely on JUCE.git being a sibling of JUCE.cmake.git on CI
Even though we clone JUCE.git as a sibling of JUCE.cmake.git on AppVeyor
and Travis, we should not assume that it will be the case when other
developers clone JUCE.cmake.git on their machine.

We have to increase the minimum CMake required version to 3.4 in order
to use the BASE_DIR subcommand of the get_filename_component() command.