Provide value for OUTPUT_NAME when $<CONFIG> is empty
When we use $<TARGET_FILE_DIR:target> in a custom command, CMake reads
the value of the OUTPUT_NAME property. If OUTPUT_NAME is not properly
defined for all possible cases, including when $<CONFIG> is empty, then
CMake will complain with the following error:

  CMake Error at JUCE.cmake/cmake/Reprojucer.cmake:1138 (add_library):
    Target 'JuceDemoPlugin_AU' OUTPUT_NAME depends on itself.
  Call Stack (most recent call first):
    examples/audio plugin demo/CMakeLists.txt:304 (jucer_project_end)

Defining OUTPUT_NAME for the case $<CONFIG:> makes it possible to use
$<TARGET_FILE_DIR:target> without having to guard it with $<CONFIG:cfg>
conditions.