Unset target name variables in order to catch target name mistakes
In the previous commit, we replaced

        __install_to_plugin_binary_location(${vst_target} "VST3"
          "$ENV{HOME}/Library/Audio/Plug-Ins/VST3"
        )

by

        __install_to_plugin_binary_location(${vst3_target} "VST3"
          "$ENV{HOME}/Library/Audio/Plug-Ins/VST3"
        )

The original code was buggy, since it was installing the wrong target in the
VST3 folder. However, CMake was still fine processing that code since
`${vst_target}` was a valid target name. When we call `unset(vst_target)`, we
ensure that subsequent occurrences of `${vst_target}` will expand to an empty
string, which is not a valid target name. Then CMake will fail if we wrongly use
`${vst_target}` instead of `${vst3_target}`.