Comment out PLATFORM_TOOLSET when translating it from .jucer
jucer_export_target() will abort if PLATFORM_TOOLSET is given but
doesn't match the toolset name given to cmake (with the option -T).
This means that Projucer 4.2.0 could only be configured with the v140_xp
(or v120_xp) toolset when using the Visual Studio 2015 (or 2013)
generator.

This is similar to how the ARCHITECTURE setting is commented out when
given to jucer_export_target_configuration() for the "Visual Studio
2015" and "Visual Studio 2013" exporters.