Fix writing audio plugin settings defines in AppConfig.h
Instead of looping over a list of setting names and values, loop over a
list of setting names only.

This fixes the issue where if a value is empty, we would treat the next
setting name as the value instead (because in most cases CMake doesn't
preserve empty items in a list).