Do not fallback to first configuration when there is only one
On macOS, the Xcode generator is the only CMake generator that supports multiple
build configurations (Debug, Release, Profile...). Other generators (Unix
Makefiles, Ninja...) support only one configuration at a time, which is chosen
by defining CMAKE_BUILD_TYPE when calling CMake. Since these generators support
only one build configuration, there is no need to loop over all available
configurations and fallback to what the first one defined. However, we keep
doing that when using the Xcode generator.

The diff of this commit is easier to read when ignoring whitespace changes. To
do so in GitHub, append "?w=1" to the URL. Setting the number of lines of
context to at least 11 also helps understanding what the added "if()" and
"else()" are surrounding.