Inline _FRUT_set_Reprojucer_current_exporter() into jucer_project_end()
Instead of aborting with "Reprojucer.cmake doesn't support any export target for
your current platform" when including Reprojucer.cmake, defer it to when calling
jucer_project_end(). This allows users of Reprojucer.cmake to include it, and
use Reprojucer_supported_exporters or Reprojucer_supported_exporters_conditions
to know whether it is fine to call jucer_project_* functions. It also allows to
include Reprojucer.cmake in test files, where none of the conditions would be
true.