Add resources.rc to all targets that need it
Because of the source files filtering (lines 1431-1451), only the *_Shared_Code
target used resources.rc, but the *_VST, *_VST3, and _*StandalonePlugin targets
should also use it.

The executable produced by the *_StandalonePlugin target now has an icon if the
ICON_SMALL and/or ICON_LARGE exporter settings have been defined.