Generate the icon file in the same folder as the .plist files
Like Projucer does.

We also change where the resources.rc is generated, since it needs to be in the
same folder as the icon file.

We still want to keep the icon file and the resources.rc file in the "Juce
Library Code" group in the IDE, however the call to source_group() with
REGULAR_EXPRESSION at line 1232 doesn't match them, so we add individual calls
to source_group().