Revert "Make it clear that we're falling back to the default value"
Having "# Default" next to some Audio Plug-in project settings can be
quite confusing. These defaults are Projucer's, not Reprojucer.cmake's.

A user of Reprojucer.cmake might be tempted to remove or comment out
"BUILD_VST ON # Default", thinking that it won't make a difference, but
it will actually stop building the VST plugin format.

This reverts commit 05d84dd08af233ab4721a83688914f7209aa8611.