Give explicit IDs to the projects created with Projucer 5.2.1 and 5.3.1
When loading a project that doesn't have an ID, Projucer generates a
random ID made of 6 alphanumerical characters. Between [1] and [2],
Projucer treated that ID has a "default value", and didn't write it in
the .jucer file when saving the project. This meant that a given project
had a different ID every time it was loaded by Projucer.

The project ID is used to compute GUIDs that are written in the build
system files. Having a different project ID means that the build system
files change when saving the project, even when the content of the
.jucer file doesn't change.

[1]: https://github.com/WeAreROLI/JUCE/commit/b72a62651
[2]: https://github.com/WeAreROLI/JUCE/commit/1e6635561