Don't try to install plugins to an empty binary destination
This can happen when "JUCER_${plugin_type}_BINARY_LOCATION_${config}"
is defined but empty, or when "default_destination" is empty, which will
be the case in the next commit.

The diff of this commit is easier to read when ignoring whitespace
changes. To do so in GitHub, append "?w=1" to the URL.