Sort private functions of Reprojucer.cmake alphabetically
The public functions of Reprojucer.cmake are sorted to match the order
in which they should be called in the CMakeLists.txt files.

The private functions of Reprojucer.cmake were kind of sorted by topic,
but it was not clear at all. Since there are 42 private functions,
sorting them alphabetically will make it way easier to navigate.

I used the Code Folding feature of Sublime Text to cut-paste the
functions without losing any line in the process.