Don't treat all files as regular source files
Files which are marked as "Xcode Resource" only or as "Xcode Resource"
and "Binary Resource" should not be treated as regular source files.

On the other hand, files which are marked as both "Compile" and "Xcode
Resource" should be treated a regular source files, otherwise they won't
be compiled.

This matches what Projucer does.