Fix double installation of bundle resources
When using `install(TARGETS)` on a target that is a bundle (i.e. it has
the MACOSX_BUNDLE or BUNDLE property set to TRUE), if that target has
files defined has resources with the RESOURCE property, these files will
also be installed in the destination.

For instance, if we had an "Audio Plug-In" project named "SuperSynth",
which had "beep.wav" as an "Xcode Resource" in Projucer, then building
(and installing) the AU target would result in both
"SuperSynth.component" and "beep.wav" to be copied in
"~/Library/Audio/Plug-Ins/Components/". Even though "beep.wav" was
already in "SuperSynth.component/Contents/Resources/" as expected.

This reverts partially what was done in
f701b0036d7894caa8fde44dc225aacb67a8da52.