Support building and installing helper executables when cross-compiling
try_compile() should be used for testing whether a project can be built
with the current CMake toolchain, not for building and installing
arbitrary software.

This change is required to support cross-compiling toolchains like iOS
and Android, since we need to build the helper executables for the host
and not the target platform.