Add --parallel to FRUT cmake build command
For better performance and experience for anyone not familiar with `cmake` (like me).
Would also add it to the build command for the project, but if Xcode is the generator then just `--parallel` doesn't work as is, must specify number of cores, I think. But let's at least accelerate FRUT's build.