Fix calling "strip" when using non-Xcode generators
Otherwise, when enabling "STRIP_LOCAL_SYMBOLS" on Release, building
in Debug using the Ninja generator will result in the following error:

  /bin/sh: : command not found

This is because CMake generates an empty post-build command instead of
no command at all, like it does when using the Xcode generator.

The diff of this commit is easier to read when ignoring whitespace
changes. To do so in GitHub, append "?w=1" to the URL.