Change the license of Reprojucer.cmake from GPLv3 to Apache 2.0
The copyright holders of Reprojucer.cmake agree to license it under the
Apache License, version 2.0, instead of under the GNU General Public
License, version 3.

This commit removes the copyright notices at the top of Reprojuce.cmake
on purpose. Each copyright holder will add back the copyright notice for
their own name in a subsequent commit where they will confirm that they
agree to the license change in the commit message.