Do not write trailing backslashes that would break CMake lists
On one hand, most jucer_* functions support taking arbitrary arguments
which are available as a semicolon-separated list in the ARGN variable.
On the other hand, backslashes can escape semicolons and change how
CMake identifies the list items.

For instance, jucer_export_target_configuration takes 5 required
arguments, so in this case:

  jucer_export_target_configuration(
    "Visual Studio 2019"
    NAME "Release"
    DEBUG_MODE OFF
    HEADER_SEARCH_PATHS
      "path\\to\\headers\\"
    LINK_TIME_OPTIMISATION ON
  )

ARGN is equal to

  HEADER_SEARCH_PATHS;path\to\headers\;LINK_TIME_OPTIMISATION;ON

but CMake will split the items as:

 - HEADER_SEARCH_PATHS
 - path\to\headers;LINK_TIME_OPTIMISATION
 - ON

because the second semicolon is escaped by the preceding backslash.

This will result in jucer_export_target_configuration complaining that
"ON" is not a known keyword.