Avoid error message from juce::LeakedObjectDetector
Leaking a few strings when exiting Jucer2CMake is not an issue, since
the OS will collect all the memory back, but that error message doesn't
look nice.

Before:

  $ ./Jucer2CMake invalid
  error: invalid mode "invalid"
  error: not enough positional arguments
  usage: Jucer2CMake {juce6,reprojucer} <jucer_project_file> [--help] [<args>]
  *** Leaked objects detected: 5 instance(s) of class StringArray
  JUCE Assertion failure in juce_LeakedObjectDetector.h:92

After:

  $ ./Jucer2CMake invalid
  error: invalid mode "invalid"
  error: not enough positional arguments
  usage: Jucer2CMake {juce6,reprojucer} <jucer_project_file> [--help] [<args>]

The diff of this commit is easier to read when ignoring whitespace
changes. To do so in GitHub, append "?w=1" to the URL.