Let code in `jucer_project_end` handle default case
We already have code that sets `architecture_flag` to `-march=native`
when `JUCER_ARCHITECTURE_FLAG_${config}` is not defined, on line 5819.
We don't need that logic to be replicated in the argument parsing code.

This change also makes it easier to change what the default arch flag is
in a later commit.