Handle new XML attributes used in `.jucer` files
Projucer now reads/writes in the `.jucer` files:
- `iosDeploymentTarget` instead of `iosCompatibility`,
- `macOSBaseSDK` instead of `osxSDK`, and
- `macOSDeploymentTarget` instead of `osxCompatibility`.

The "macOS Base SDK Version" configuration setting was also renamed to
"macOS Base SDK".

This matches what was done in
https://github.com/juce-framework/JUCE/commit/9dc98fb1d, which was
released in JUCE 6.1.0.

The diff of this commit is easier to read when ignoring whitespace
changes. To do so in GitHub, append "?w=1" to the URL.