Prevent CMake from setting `ARCHS` to `$(NATIVE_ARCH_ACTUAL)`
On one hand, by default, Projucer doesn't write the `ARCHS` attribute in
the generated Xcode project, which is interpreted by Xcode 12 and later
as `$(ARCHS_STANDARD)`.

On the other hand, since [1] (which was released in CMake 3.18.1), by
default, CMake sets the `ARCHS` attribute in the generated Xcode
project to `$(NATIVE_ARCH_ACTUAL)`.

This means that, by default, an Xcode project generated by Projucer will
build for several architecture, while an Xcode project generated by
CMake will only build for the host architecture.

Since Reprojucer.cmake tries to match the behavior of Projucer as
closely as possible, we have to "undo" what CMake tries to do and set
`ARCHS` to `$(ARCHS_STANDARD)` explicitly.

[1]: https://gitlab.kitware.com/cmake/cmake/-/commit/26673bf48011a48fe0ad01e97409a32c1f50143d